+++ sys/compat/linuxkpi/common/include/linux/scatterlist.h
@@ -43,7 +43,7 @@
 
 struct bus_dmamap;
 struct scatterlist {
-	unsigned long page_link;
+	uintptr_t page_link;
 #define	SG_PAGE_LINK_CHAIN	0x1UL
 #define	SG_PAGE_LINK_LAST	0x2UL
 #define	SG_PAGE_LINK_MASK	0x3UL
@@ -120,7 +120,7 @@ sg_assign_page(struct scatterlist *sg, struct page *page)
 {
 	unsigned long page_link = sg->page_link & SG_PAGE_LINK_MASK;
 
-	sg->page_link = page_link | (unsigned long)page;
+	sg->page_link = page_link | (uintptr_t)page;
 }
 
 static inline void
@@ -166,7 +166,7 @@ static inline void *
 sg_virt(struct scatterlist *sg)
 {
 
-	return ((void *)((unsigned long)page_address(sg_page(sg)) + sg->offset));
+	return ((void *)((uintptr_t)page_address(sg_page(sg)) + sg->offset));
 }
 
 static inline void
@@ -177,7 +177,7 @@ sg_chain(struct scatterlist *prv, unsigned int prv_nents,
 
 	sg->offset = 0;
 	sg->length = 0;
-	sg->page_link = ((unsigned long)sgl |
+	sg->page_link = ((uintptr_t)sgl |
 	    SG_PAGE_LINK_CHAIN) & ~SG_PAGE_LINK_LAST;
 }
 
@@ -215,7 +215,7 @@ static inline void
 sg_kfree(struct scatterlist *sg, unsigned int nents)
 {
 	if (nents == SG_MAX_SINGLE_ALLOC) {
-		free_page((unsigned long)sg);
+		free_page((uintptr_t)sg);
 	} else
 		kfree(sg);
 }
