+++ sys/compat/linuxkpi/common/include/linux/xarray.h
@@ -126,7 +126,7 @@ xa_init(struct xarray *xa)
 static inline void *
 xa_mk_value(unsigned long v)
 {
-	unsigned long r = (v << 1) | 1;
+	uintptr_t r = (v << 1) | 1;
 
 	return ((void *)r);
 }
@@ -134,7 +134,7 @@ xa_mk_value(unsigned long v)
 static inline bool
 xa_is_value(const void *e)
 {
-	unsigned long v = (unsigned long)e;
+	uintptr_t v = (uintptr_t)e;
 
 	return (v & 1);
 }
@@ -142,7 +142,7 @@ xa_is_value(const void *e)
 static inline unsigned long
 xa_to_value(const void *e)
 {
-	unsigned long v = (unsigned long)e;
+	uintptr_t v = (uintptr_t)e;
 
 	return (v >> 1);
 }
